Feature/ support aggregate-over and non-aggregate-over in streaming queries (#1179)
* support aggregate-over and non-aggregate-over in streaming queries

* refactor substream

* support partition by syntax

* add smoke test

+++ src/NativeLog/Log/LogManagerConfig.h
@@ -10,7 +10,7 @@ namespace nlog
 struct LogManagerConfig
 {
     uint32_t recovery_threads_per_data_dir = 1;
-    uint32_t flush_check_ms = std::numeric_limits<int64_t>::max();
+    uint32_t flush_check_ms = std::numeric_limits<uint32_t>::max();
     uint32_t flush_recovery_sn_checkpoint_ms = 60000;
     uint32_t flush_start_sn_checkpoint_ms = 60000;
     uint32_t retention_check_ms = 5 * 60 * 1000;
